+++ packages/orq-rc/src/orq_ai_sdk/models/filegetop.py
@@ -60,5 +60,5 @@ class FileGetResponseBody(BaseModel):
     workspace_id: str
     r"""The id of the resource"""
 
-    created: Optional[datetime] = dateutil.parser.isoparse("2025-03-07T11:47:15.597Z")
+    created: Optional[datetime] = dateutil.parser.isoparse("2025-03-10T09:56:39.106Z")
     r"""The date and time the resource was created"""